Add Domain models, small fixes

+++ back/Cumpuss.Database/dbo/Classes.sql
@@ -3,7 +3,7 @@
 	[Id] INT NOT NULL PRIMARY KEY, 
     [CourseId] INT NOT NULL,
     [Date] DATE NOT NULL, 
-    [ClassType] SMALLINT NOT NULL, 
+    [ClassType] INT NOT NULL, 
     [CustomCaption] NVARCHAR(20) NULL, 
     CONSTRAINT [FK_Classes_ToCourses] FOREIGN KEY ([CourseId]) REFERENCES [Courses]
 )
